+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * added `setDeprecated()` method to indicate a deprecated node
  * added `XmlUtils::parse()` method to parse an XML string
+ * deprecated `ConfigCachePass`
 
 3.3.0
 -----